Move normandy serializers
