+++ .gitignore
@@ -25,10 +25,6 @@ var/
 .installed.cfg
 *.egg
 
-#Symlink
-normandy
-delivery-console
-
 # PyInstaller
 #  Usually these files are written by a python script from a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
