+++ app/experimenter/experiments/api_views.py
@@ -25,7 +25,7 @@ from experimenter.experiments.serializers.design import (
 from experimenter.experiments.serializers.timeline_population import (
     ExperimentTimelinePopSerializer,
 )
-from experimenter.experiments.serializers.recipe import ExperimentRecipeSerializer
+from experimenter.normandy.serializers import ExperimentRecipeSerializer
 
 
 class ExperimentListView(ListAPIView):
