+++ app/experimenter/experiments/models.py
@@ -379,7 +379,7 @@ class Experiment(ExperimentConstants, models.Model):
 
     @property
     def normandy_recipe_json(self):
-        from experimenter.experiments.serializers.recipe import ExperimentRecipeSerializer
+        from experimenter.normandy.serializers import ExperimentRecipeSerializer
 
         return json.dumps(ExperimentRecipeSerializer(self).data, indent=2)
